no notification in case no selection

+++ app/src/main/java/com/github/autodiag2/elm327emu/sim.kt
@@ -83,6 +83,9 @@ class ConfigAdapter(
     }
 
     fun onDelete() {
+        if ( selectedItems.firstOrNull() == null ) {
+            return
+        }
         for(config in selectedItems) {
             config.file.delete()
         }
